backend/requirements.txt
# Python dependencies
fastapi
uvicorn
openai
python-multipart
sqlalchemy
python-dotenv
